Made the posibililty to view the reservations as a User. And Cancel it if the User wants to do it.
I also fixed a Update Bug. I tried to solve it with code, but didn't work so I went for investigation and found the solution for the Update Bug. If the application user is using a Dutch windows language system. The data notation won't work well with Visual Studio.

+++ DTO layer/Entities DTO/VehicleDTO.cs
@@ -10,7 +10,7 @@ namespace DTO_layer
         public string Transmission { get; set; }
         public int Enginepower { get; set; }
         public int Weight { get; set; }
-        public double Acceleration { get; set; }
+        public decimal Acceleration { get; set; }
         public int Cargospace { get; set; }
         public int Seat { get; set; }
         public double RentalPrice { get; set; }
