+++ Fontys S2 Project - Car to go_BLL/Logic interfaces/IVehicle.cs
@@ -12,7 +12,7 @@ namespace Logic_interfaces
         public string Modelname { get; set; }
         public string Transmission { get; set; }
         public int Enginepower { get; set; }
-        public double Acceleration { get; set; }
+        public decimal Acceleration { get; set; }
         public int Weight { get; set; }
         public int Cargospace { get; set; }
         public int Seat { get; set; }
