+++ Fontys S2 Project - Car to go_BLL/Models/Vehicle.cs
@@ -14,7 +14,7 @@ namespace BLL.Models
         public string Modelname { get; set; }
         public string Transmission { get; set; }
         public int Enginepower { get; set; }
-        public double Acceleration { get; set; }
+        public decimal Acceleration { get; set; }
         public int Weight { get; set; }
         public int Cargospace { get; set; }
         public int Seat { get; set; }
